+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [Unreleased]
+
+### Added
+
+- `Formal\AccessLayer\Query\Parameter\Type::for()` to determine a type for any given value
+
 ## 2.15.0 - 2024-02-10
 
 ### Added